move deadlock detection to its own folder
